requirements.txt
aiohttp==3.7.3
aiosqlite==0.17.0
alabaster==0.7.12
arrow==0.17.0
asgiref==3.3.1
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.9.0
braceexpand==0.1.6
cachetools==4.2.1
cassiopeia==4.0.14
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
Cogs==0.4.4
colorama==0.4.4
datapipelines==1.0.7
dblpy==0.4.0
decorator==4.4.2
discord.py==1.6.0
Django==3.1.7
django-recaptcha==2.0.6
docutils==0.16
Flask==1.1.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httplib2==0.19.0
httpx==0.13.3
humanize==3.2.0
hyperframe==5.2.0
idna==2.10
imagesize==1.2.0
import-expression==1.1.4
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jishaku==1.20.0.220
MarkupSafe==1.1.1
merakicommons==1.0.9
multidict==4.7.6
networkx==2.5
packaging==20.9
parsedatetime==2.6
Pillow==8.1.0
ping3==2.6.6
pluggy==0.13.1
praw==7.1.0
prawcore==1.5.0
protobuf==3.14.0
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pymongo==3.11.3
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
ratelimiter==1.2.0.post0
requests==2.25.1
rfc3986==1.4.0
rotate-screen==0.1.0
rsa==4.7
rule34==1.7.5
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.1.0
speedtest==0.0.1
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.26.2
wavelink==0.9.6
websocket-client==0.57.0
Werkzeug==1.0.1
yagmail==0.11.224
yarl==1.6.0
youtube-dl==2021.1.24.1
zipp==3.4.0
